Rename + add documentation!
